Rename wallet context to transactions
