+++ src/pages/Wallet/WalletHomePage.tsx
@@ -32,7 +32,7 @@ import { FloatingPanel, Section } from '../../components/PageComponents/PageCont
 import Spinner from '../../components/Spinner'
 import TransactionItem from '../../components/TransactionItem'
 import { useGlobalContext } from '../../contexts/global'
-import { SimpleTx, useWalletContext } from '../../contexts/wallet'
+import { SimpleTx, useTransactionsContext } from '../../contexts/transactions'
 import { ReactComponent as AlephiumLogoSVG } from '../../images/alephium_logo_monochrome.svg'
 import { appHeaderHeight, deviceBreakPoints } from '../../style/globalStyles'
 import { getHumanReadableError } from '../../utils/api'
@@ -46,7 +46,7 @@ const WalletHomePage = () => {
   const history = useHistory()
   const { wallet, setSnackbarMessage, client, lockWallet, currentUsername, currentNetwork } = useGlobalContext()
   const [balance, setBalance] = useState<bigint | undefined>(undefined)
-  const { networkPendingTxLists, loadedTxList, setLoadedTxList } = useWalletContext()
+  const { networkPendingTxLists, loadedTxList, setLoadedTxList } = useTransactionsContext()
   const [totalNumberOfTx, setTotalNumberOfTx] = useState(0)
   const [isLoading, setIsLoading] = useState(false)
   const [isHeaderCompact, setIsHeaderCompact] = useState(false)
